Generalized Boxplot

This function computes the Generalized Boxplot, a robust data visualization technique for detecting outliers and skewness in a distribution, as proposed by Bruffaerts et al. (2014). It uses the Tukey g-and-h distribution (from the package TukeyGH77) to model the data.

Usage

generalized_boxplot(x, alpha = 0.95, gamma = 0.5)

Arguments

x

A numeric vector containing the data.

alpha

A numeric value between 0 and 1 specifying the desired coverage rate. Default is 0.95.

gamma

gamma to be defined

Value

A list containing the following components:

lower.fence

The lower fence value.

lower.quartile

The lower quartile value.

median

The median value.

upper.quartile

The upper quartile value.

upper.fence

The upper fence value.

lower.outliers

A vector containing the lower outliers.

upper.outliers

A vector containing the upper outliers.

References

Hubert, M., and Vandervieren, E. (2008). An adjusted boxplot for skewed distributions. Computational Statistics & Data Analysis, 52(12):5186-5201.

Bruffaerts, C., Verardi, V., Vermandele, C. (2014). A generalized boxplot for skewed and heavy-tailed distributions. Statistics and Probability Letters 95(C):110–117

Author

Christian L. Goueguel

Examples

# Compute the Generalized Boxplot for a normal distribution
x <- rnorm(100)
gb <- generalized_boxplot(x)
print(gb)
#> $lower.fence
#> [1] 0.2073812
#> 
#> $lower.quartile
#> [1] -0.3043328
#> 
#> $median
#> [1] 0.213958
#> 
#> $upper.quartile
#> [1] 0.7480812
#> 
#> $upper.fence
#> [1] 0.213958
#> 
#> $lower.outliers
#>  [1] -2.210633111 -2.078489269 -2.037682494 -1.943650901 -1.728304515
#>  [6] -1.692101521 -1.667928046 -1.617773765 -1.614039457 -1.547776544
#> [11] -1.540442405 -1.529310531 -1.364709458 -1.318220727 -1.176692158
#> [16] -1.041673294 -0.904215026 -0.788963218 -0.732853806 -0.728219111
#> [21] -0.693094614 -0.482676822 -0.436829977 -0.414588726 -0.321892586
#> [26] -0.321324842 -0.287340800 -0.279072171 -0.278454025 -0.239029068
#> [31] -0.224961271 -0.212848279 -0.201458147 -0.195889249 -0.186120699
#> [36] -0.162421942 -0.120393825 -0.109710321 -0.097451250 -0.096686073
#> [41] -0.093636794 -0.091434276 -0.086714135 -0.075263198 -0.056055946
#> [46] -0.004060653  0.001683688  0.002691661  0.118849433  0.189426238
#> 
#> $upper.outliers
#>  [1] 0.2263249 0.2502478 0.2796279 0.2893440 0.3687421 0.3732414 0.3746436
#>  [8] 0.4032903 0.4205742 0.4574621 0.4749117 0.4836175 0.5114708 0.5249143
#> [15] 0.5397906 0.5638674 0.5899827 0.6164557 0.6165678 0.6220332 0.6418300
#> [22] 0.6644159 0.7116019 0.7480812 0.8134004 0.8344016 0.8811788 0.9044355
#> [29] 0.9678592 0.9863659 0.9902622 1.0128343 1.0191571 1.0610953 1.1250719
#> [36] 1.1871868 1.1916013 1.2765787 1.3995762 1.4414618 1.4878378 1.5384302
#> [43] 1.7386338 1.8778640 2.3747347 2.3829267 2.4580605 2.6917140
#> 

# Compute the Generalized Boxplot for a skewed distribution
x <- rexp(100)
gb <- generalized_boxplot(x)
print(gb)
#> $lower.fence
#> [1] 0.624786
#> 
#> $lower.quartile
#> [1] 0.2687795
#> 
#> $median
#> [1] 0.6581434
#> 
#> $upper.quartile
#> [1] 1.27906
#> 
#> $upper.fence
#> [1] 0.6581434
#> 
#> $lower.outliers
#>  [1] 0.02532230 0.02576994 0.04289057 0.04482934 0.05122371 0.05773039
#>  [7] 0.06910619 0.07610867 0.08872914 0.11328210 0.11663788 0.11762548
#> [13] 0.12365090 0.12954107 0.18277448 0.19361383 0.20726172 0.21769428
#> [19] 0.21812196 0.22349682 0.23512627 0.25149112 0.25500488 0.26114497
#> [25] 0.26687317 0.27068574 0.27837162 0.28413990 0.30876258 0.31117094
#> [31] 0.31299193 0.32367034 0.33836365 0.34941053 0.41094150 0.42606612
#> [37] 0.43237343 0.44444410 0.48270025 0.48417332 0.49056389 0.49218720
#> [43] 0.54209259 0.55987508 0.57015672 0.57036588 0.58425699 0.61869255
#> 
#> $upper.outliers
#>  [1] 0.6973561 0.7149305 0.7231569 0.7328592 0.7511337 0.7854578 0.7995130
#>  [8] 0.8009929 0.8298000 0.8327294 0.8388340 0.8686520 0.9348019 0.9486319
#> [15] 0.9864915 1.0044356 1.0088133 1.0750821 1.0794319 1.1183679 1.2025156
#> [22] 1.2239947 1.2567066 1.2662859 1.2790601 1.2933980 1.3756673 1.3852603
#> [29] 1.7599141 1.8129885 1.8883480 1.9321722 2.0381937 2.0987510 2.1112117
#> [36] 2.1311824 2.1553678 2.2362940 2.2904714 2.3969276 2.4405611 2.4917445
#> [43] 2.7955856 2.9029413 3.0515370 3.0977053 3.6436161 3.7143403 3.9938567
#> [50] 4.0723257
#>