Robust Box-Cox and Yeo-Johnson Transformation

Transforms each variable in a dataset toward central normality using re-weighted maximum likelihood to robustly fit the Box-Cox or Yeo-Johnson transformation.

Usage

robTransform(.data, var = NULL, type = "bestObj", quant = 0.99, nbsteps = 2)

Source

Raymaekers and Rousseeuw (2021), Machine Learning, https://doi.org/10.1007/s10994-021-05960-5.

Arguments

.data

data frame or tibble.

var

vector of selected character or numeric variable(s). Default is NULL (all columns are selected).

type

transformation method to use. Available methods are "BC" (Box-Cox power transformation, which is for strictly positive values.), "YJ" ( Yeo-Johnson power transformation, which works for positive and negative values.) or "bestObj" (both BC and YJ are run for strictly positive variables, and the solution with lowest objective is kept. Whereas if a variable has negative values YJ is run.). Default is "bestObj".

quant

numeric value for the quantile used in determining the weights in the re-weighting step (default is 0.99).

nbsteps

number of re-weighting steps (2 by default).

Value

tibble of transformed variable(s), method used ('BC' for Box-Cox and 'YJ' for Yeo-Johnson), objective and lambda.

Details

Wrapper of the transfo function from the cellWise package.

Author

Christian L. Goueguel