Deprecated functions in specProc

These functions were renamed in specProc 0.2.0 to follow a consistent snake_case naming scheme. The old names still work but issue a deprecation warning and will be removed in a future release.

Usage

whittaker(...)

lorentzian(...)

pseudo_voigt(...)

peakfit(...)

multipeakfit(...)

plotfit(...)

plotSpec(...)

outlierplot(...)

directOutlyingness(...)

iqrMethod(...)

robustBCYJ(...)

rousseeuwCroux(...)

summaryStats(...)

tukeyGH(...)

yGradientglsw(...)

pareto(...)

Arguments

...

Arguments passed to the new function.

Value

The result of the new function.

Details

gaussian() was renamed to gaussian_profile() without an alias, because an exported gaussian() masks stats::gaussian(), the family used by glm().

Old name	New name

whittaker()	baseline_als()

lorentzian()	lorentzian_profile()

pseudo_voigt()	pseudo_voigt_profile()

peakfit()	peak_fit()

multipeakfit()	multipeak_fit()

plotfit()	plot_fit()

plotSpec()	plot_spectra()

outlierplot()	plot_outliers()

directOutlyingness()	directional_outlyingness()

iqrMethod()	iqr_outliers()

robustBCYJ()	robust_bcyj()

rousseeuwCroux()	rousseeuw_croux()

summaryStats()	summary_stats()

tukeyGH()	tukey_gh()

yGradientglsw()	y_gradient_glsw()

pareto()	pareto_scale()