Modified Orthogonal Projections to Latent Structures

This function implements the modified orthogonal projections to latent structures (O2PLS) algorithm as proposed by Trygg (2002). The OPLS and O2PLS methods differ as follows: OPLS is unidirectional \((X \Rightarrow Y)\), meaning that only orthogonal variations in the \(X\)-space are filtered out. Whilst O2PLS is bi-directional \((X \Leftrightarrow Y)\), meaning that orthogonal variations in both the \(X\)- and \(Y\)-space are filtered out.

Usage

o2pls(x, y, ncomp = 10, center = TRUE, scale = FALSE, tol = 1e-05)

Arguments

x

A numeric matrix or data frame representing the predictor variables.

y

A numeric vector, matrix or data frame representing the response variables.

ncomp

An integer representing the number of components. Default value is 10.

center

A logical value indicating whether to mean-centered x and y. Default is TRUE.

scale

A logical value indicating whether to scale x and y. Default is FALSE.

tol

A numeric value representing the tolerance for convergence. The default value is 1e-5.

Value

A list containing the following components:

correction: The corrected matrix.

scores: The orthogonal scores matrix.

loadings: The orthogonal loadings matrix.

weights: The weights matrix.

Details

The O2PLS method handles situations where systematic \(X\)-orthogonal variation in \(Y\) exists, and it is predictive in both ways, \((X \Rightarrow Y)\) and \((Y \Rightarrow X)\). O2PLS uses least squares regression to estimate the pure constituent profiles and divide the systematic part in \(X\) and \(Y\) into two parts, one which is related to both \(X\) and \(Y\) (covarying) and one that is not (orthogonal).

References

Trygg, J., (2002). O2-PLS for qualitative and quantitative analysis in multivariate calibration. J. Chemom. 16(1):283–293.

Author

Christian L. Goueguel