Orthogonal Signal Correction

Orthogonal signal correction (OSC) removes systematic variation in the spectral matrix that is orthogonal to the response vector.

Usage

osc(
  x,
  y,
  method = "wold",
  center = TRUE,
  osc.ncomp = 4,
  pls.ncomp = 10,
  tol = 0.001,
  iter = 20
)

Arguments

x

A matrix or data.frame containing the x-data (predictors).

y

A vector containing the y-data (response). The length of y must match the number of rows in x.

method

A character string indicating the OSC method to use. Accepted values are "wold" (Wold's original method), "sjoblom" (Sjöblom's method), or "wise" (Wise and Gallagher's method).

center

A logical value indicating whether to center the x-data before applying OSC. If TRUE, the columns of x will be mean-centered.

osc.ncomp

An integer representing the number of OSC components to be calculated. The default value is 4.

pls.ncomp

An integer representing the number of PLS components to be calculated. The default value is 10.

tol

A numeric value representing the tolerance for convergence of the OSC algorithm. The default value is 1e-3.

iter

An integer representing the maximum number of iterations for the OSC algorithm. The default value is 20.

Value

An list containing the following components:

x_corrected

The corrected x-data after applying OSC.

scores

The x-scores after orthogonal variations are removed.

loadings

The loadings.

weights

The wights.

R2

The value of R2

Details

The OSC algorithm identifies and removes the orthogonal variation in the x-data by iteratively deflating the x-matrix with respect to the y-data. The resulting x-matrix contains only the variation that is relevant to the y-data, which can then be used for further modeling or analysis. This function implements three different methods for OSC: the original method proposed by Wold et al. (1998), the method proposed by Sjöblom et al. (1998), and the method proposed by Wise and Gallagher.

References

Sjoblom, J., Svensson, O., Josefson, M., Kullberg, H., Wold, S. (1998). An evaluation of orthogonal signal correction applied to calibration transfer of near infrared spectra. Chemometrics Intell. Lab. Syst., 44(1):229-244.

Wise, B. M. and Gallagher, N.B. http://www.eigenvector.com/MATLAB/OSC.html.

Wold, S., Antti, H., Lindgren, F., Ohman, J. (1998). Orthogonal signal correction of near-infrared spectra. Chemometrics Intell. Lab. Syst., 44(1):175-185.

Svensson, O., Kourti, T. and MacGregor, J.F. (2002). An investigation of orthogonal correction algorithms and their characteristics. Journal of Chemometrics, 16(1):176-188.

Westerhuis, J. A., de Jong, S., Smilde, A. K. (2001). Direct orthogonal signal correction. Chemometrics Intell. Lab. Syst., 56(1):13-25.

Author

Christian L. Goueguel

Examples

if (FALSE) {
osc_result <- osc(
  x_data, y_data,
  method = "wold",
  center = TRUE,
  osc.ncomp = 4,
  pls.ncomp = 10
)
}