Piecewise Direct Standardization

The pds function performs Piecewise Direct Standardization (PDS), a method proposed by Bouveresse & Massart (1996) to transfer spectra from one instrument to another. The main improvement over the original PDS method, Wang et al., (1991), is the use of a global PLSR model in addition to the local PLSR models, which can lead to better transfer performance.

Usage

pds(x1, x2, win, ncomp, alpha)

Arguments

x1

A matrix or data frame containing spectra acquired with the standard instrument.

x2

A matrix or data frame containing spectra acquired with the instrument to be standardized.

win

An integer specifying the half size of the moving window used for PLSR. A larger value may improve the transfer but will increase computational time.

ncomp

An integer specifying the number of components to be used in PLSR. Typically, a small number (e.g., 2-5) is sufficient.

alpha

A numeric value between 0 and 1 specifying the weight for the global PLSR model. A value of 0 corresponds to the original PDS method, while a value of 1 corresponds to using only the global PLSR model.

Value

A list with two components:

transfert_matrix:

A matrix containing the transfer coefficients.

intercept:

A vector containing the intercepts for each local PLSR model.

References

Wang, Y., Veltkamp, D.J., Kowalski, B.R., (1991). Multivariate instrument standardization. Analytical Chemistry, 63(23):2750-2756.

Bouveresse, E., Massart, D.L., (1996). Improvement of the piecewise direct standardization procedure for the transfer of NIR spectra for multivariate calibration. Chemometrics and Intelligent Laboratory Systems, (32)2:201-213.