y-Gradient Generalized Least Squares Weighting

The y-gradient generalized least squares weighting algorithm (GLSW) removes variance from the data (spectra), which is orthogonal to the response.

Usage

yGradientglsw(x, y, alpha = 0.01)

Arguments

x

A numeric matrix, data frame or tibble, representing the predictors data.

y

A numeric vector representing the response vector.

alpha

A numeric value specifying the weighting parameter. Typical values range from 1 to 0.0001. Default is 0.01.

Value

A tibble containing the filtering matrix.

Details

The y-Gradient GLSW is an alternative method to GLSW, where a continuous \(\textbf{y}\)-variable is used to develop pseudo-groupings of samples in \(\textbf{X}\) by comparing the differences in the corresponding \(\textbf{y}\) values. This is referred to as the "gradient method" because it utilizes a gradient of the sorted \(\textbf{X}\)- and \(\textbf{y}\)-blocks to calculate a covariance matrix.

References

Zorzetti, B.M., Shaver, J.M., Harynuk, J.J., (2011). Estimation of the age of a weathered mixture of volatile organic compounds. Analytica Chimica Acta, 694(1-2):31–37.