Classical or Robust Z-Score

This function is a wrapper around the robustHD package and calculates classical or robust z-score (standardization) for a numeric vector.

Usage

zscore(x, robust = FALSE)

Arguments

x

A numeric vector.

robust

A logical value indicating whether to calculate classical (default) or robust z-score.

Value

A tibble with the the input numeric vector and the calculated z-scores arranged in descending order.

Details

Z-scores are useful for comparing data points from different distributions because they are dimensionless and standardized. A positive z-score indicates that the data point is above the mean, while a negative z-score indicates that the data point is below the mean. One common rule to detect outliers using z-scores is the "three-sigma rule", in which data points with an absolute z-score greater than 3 (|z| > 3) can be considered potential outliers, as they fall outside the range that covers 99.7% of the data points in a normal distribution. (Note that a cutoff of |z| > 2.5 is also often used).

Rousseeuw, P. J., & Croux, C. (1993). Alternatives to the median absolute deviation. Journal of the American Statistical Association, 88(424), 1273-1283.

Rousseeuw, P. J., & Hubert, M. (2011). Robust statistics for outlier detection. Wiley Interdisciplinary Reviews: Data Mining and Knowledge Discovery, 1(1), 73-79.

Author

Christian L. Goueguel

Examples

vec <- c(1, 2, 3, 4, 5, 100)
zscore(vec)
#> # A tibble: 6 × 2
#>    data  score
#>   <dbl>  <dbl>
#> 1   100  2.04 
#> 2     5 -0.358
#> 3     4 -0.383
#> 4     3 -0.408
#> 5     2 -0.433
#> 6     1 -0.458
zscore(vec, robust = TRUE)
#> # A tibble: 6 × 2
#>    data  score
#>   <dbl>  <dbl>
#> 1   100 43.4  
#> 2     5  0.674
#> 3     4  0.225
#> 4     3 -0.225
#> 5     2 -0.674
#> 6     1 -1.12